move qr code scanner from ContactSelector to ContactSelectorOptions
